Moved token actions to separate actions class
