+++ .env.sample
@@ -9,3 +9,6 @@ REACTIONS_PREFIX=reactions
 
 # Postfix for tokes collection
 TOKENS_POSTFIX=tokens
+
+# Indicates how long token is valid
+TOKEN_LIFETIME_IN_MINUTES=10
