--- src/constants/token.ts
@@ -1,1 +1,0 @@
-export const TOKEN_LIFETIME_IN_MINUTES = 10;
